requirements.txt
streamlit
torch
transformers
fire
Pillow
einops
streamlit-keyup